+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - rpc: change upstream header to `x-bigtable: disabled` ([#5](https://github.com/lamports-dev/alpamayo/pull/5))
 - rpc: use confirmed during sync ([#6](https://github.com/lamports-dev/alpamayo/pull/6))
 - storage: remove transaction index on slot remove ([#8](https://github.com/lamports-dev/alpamayo/pull/8))
+- storage: set confirmed/finalized on first stream messages ([#17](https://github.com/lamports-dev/alpamayo/pull/17))
 
 ### Features